requirements.txt
# Core
numpy==2.2.6
polars==1.34.0
pyarrow==21.0.0
PyYAML==6.0.3
tqdm==4.66.5

# PyTorch stack
torch==2.8.0
torchaudio==2.9.0

# Logging
tensorboard==2.18.0

# HF / embeddings (Amazon preprocessing)
datasets==4.2.0
huggingface-hub==0.35.3
transformers==4.57.1
tokenizers==0.22.1
safetensors==0.6.2
sentence-transformers==5.2.2

# Optional acceleration (comment out if install fails)
triton==3.4.0
flash_attn==2.8.3
xformers==0.0.32.post2

# ANN search (if you use it in retrieval / eval)
faiss-gpu-cu12==1.13.0

# Experiment tracking (optional)
wandb==0.24.0
